+++ admin.ts
@@ -1,3 +1,4 @@
+import {Client} from "./client";
 
 export interface Log {
     timestamp: Date;
@@ -11,3 +12,12 @@ export enum AdminPrivileges {
     MANAGER = 0,
     SUPERUSER = 1,
 }
+
+export interface NewClient {
+    name: string;
+    primary_email: string;
+    licenses: number;
+    containers: string[];
+}
+
+export type ClientUpdate = Partial<Omit<Client, "id" | "created_at">> & { id: Client["id"] };
